fix: es client search definition
